split side_t in static and dynamic data

+++ p_spec.h
@@ -10,7 +10,7 @@
  *  Jess Haas, Nicolas Kalkhof, Colin Phipps, Florian Schulze
  *  Copyright 2005, 2006 by
  *  Florian Schulze, Colin Phipps, Neil Stevens, Andrey Budko
- *  Copyright 2023, 2024 by
+ *  Copyright 2023-2026 by
  *  Frenkel Smeijers
  *
  *  This program is free software; you can redistribute it and/or
@@ -164,7 +164,7 @@ typedef struct
 {
   const line_t __far* line;
   bwhere_e where;
-  int16_t   btexture;
+  int8_t   btexture;
   int16_t   btimer;
   degenmobj_t __far* soundorg;
